requirements.txt
APScheduler==3.7.0
boto==2.49.0
boto3==1.16.63
botocore==1.19.63
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==2.2.3
django-allauth==0.44.0
django-apscheduler==0.5.2
django-cors-headers==3.0.2
django-heroku==0.3.1
django-s3direct==1.1.8
django-static-fontawesome==5.14.0.0
django-storages==1.11.1
djangorestframework==3.10.2
gunicorn==19.9.0
idna==2.10
jmespath==0.10.0
mailchimp-marketing==3.0.55
oauthlib==3.1.0
Pillow==7.0.0
psycopg2==2.8.3
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.17.1
python3-openid==3.2.0
pytz==2019.1
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.3.7
schedule==0.6.0
six==1.15.0
sqlparse==0.3.0
stripe==2.56.0
tzlocal==2.1
urllib3==1.26.4
whitenoise==4.1.3
